feat: new dogs banner
